only show Sync config when that feature is supported
